Penambahan behaviour dengan menggunakan Mixin pada class

+++ test/animal.dart
@@ -2,16 +2,19 @@ abstract class Animal {
   String? name;
   String? category;
   String? habitat;
-  late int? _weight;
+  late double? _weight;
 
   Animal(
-      {required this.name, required int weight, this.category, this.habitat}) {
+      {required this.name,
+      required double weight,
+      this.category,
+      this.habitat}) {
     _weight = weight;
   }
 
-  int get weight => _weight!;
+  double get weight => _weight!;
 
-  int eating(int foodWeight) {
+  double eating(int foodWeight) {
     return _weight = _weight! + foodWeight;
   }
 }
